Fixes channel going back in time due to incomplete change.
The change in d3d33d5b2518e4e6516b6f31270b406937c48ca2 changed how the
locally saved channel information was saved on the local machine.

 * https://github.com/NixOS/nixos-channel-scripts/commit/d3d33d5b2518e4e6516b6f31270b406937c48ca2#diff-682bf482f6dd273f1c3c49afd99b4660R192

The changeset did *not* change the guard clause, which now made it act
like this:

 1. Try readlink on the file
 2. (bail as it's undefined)

The new behaviour is as more or less expected:

 1. Reads the file if available
 2. Split the URL components
 3. Keep the last bit

This can give us an empty string, which will compare to `-1` thus
fulfilling the need.